refactor(motion): use local reduced motion hook
