+++ packages/ibm-products/src/components/NotificationsPanel/NotificationsPanel.tsx
@@ -38,7 +38,7 @@ import {
   useIsomorphicEffect,
   usePreviousValue,
 } from '../../global/js/hooks';
-import usePrefersReducedMotion from '../../global/js/hooks/usePrefersReducedMotion';
+import { usePrefersReducedMotion } from '../../global/js/hooks';
 import wrapFocus from '../../global/js/utils/wrapFocus';
 
 // The block part of our conventional BEM class names (blockClass__E--M).
