+++ packages/ibm-products/src/global/js/hooks/index.js
@@ -19,3 +19,4 @@ export { useControllableState } from './useControllableState';
 export { usePrefix } from './usePrefix';
 export { useFocus } from './useFocus';
 export { useIsomorphicEffect } from './useIsomorphicEffect';
+export { usePrefersReducedMotion } from './usePrefersReducedMotion';
